Blood Splash/Puddle, Swinging Axe Trap, Calm Bonus, bugfixes
Blood Splash/Puddle - VFX activated when an agent dies

Swinging Axe Trap - Trap that is randomly placed in the game level at the beginning of the round, intended to kill any agents that hits it. !!! Agents can see and avoid this type of trap in their own !!!

One For All and All For One -  when an agent finishes the round it applies a 10p calm to every alive agent. !!! Saboteur agent type doesn't receive this bonus !!!

Minor bugfixes.